+++ scripts/pf2e-kineticists-companion.js
@@ -1,6 +1,8 @@
+import { Settings } from "./config/settings.js";
 import { ArmorInEarth } from "./actions/armor-in-earth.js";
 import { FreshProduce } from "./actions/fresh-produce.js";
 import { ThermalNimbus } from "./actions/thermal-nimbus.js";
+import { MetalCarapace } from "./actions/metal-carapace.js";
 
 export let DamageRoll = null;
 
@@ -9,8 +11,10 @@ Hooks.on(
     () => {
         DamageRoll = CONFIG.Dice.rolls.find((r) => r.name === "DamageRoll");
 
+        Settings.initialise();
         ThermalNimbus.initialise();
         FreshProduce.initialise();
         ArmorInEarth.initialise();
+        MetalCarapace.initialise();
     }
 );